+++ packages/drupal-kit/src/fetch/defaultFetch.ts
@@ -2,28 +2,39 @@ import { ServerResponse } from 'http';
 import fetch from 'isomorphic-fetch';
 
 import addSurrogateKeyHeader from '../utils/addSurrogateKeyHeader';
-import updateMaxAge from '../utils/updateMaxAge';
+
+const defaultCacheControlValue =
+  'public, s-maxage=10, stale-while-revalidate=600';
+
 /**
  * fetch data from a JSON:API endpoint, bubbling up surrogate keys if possible
  * @param apiUrl the api url for the JSON:API endpoint
  * @param requestInit fetch initialization object
  * @param res response object
+ * @param cacheControl optional value to override cache control header, defaults to 'public, s-maxage=10, stale-while-revalidate=600'
  * @returns a promise containing the data for the JSON:API response
  */
 const defaultFetch = (
   apiUrl: RequestInfo,
   requestInit: RequestInit = {},
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  res?: ServerResponse | boolean
+  res?: ServerResponse | boolean,
+  cacheControl: string = defaultCacheControlValue
 ): Promise<Response> => {
   // Use the existing request headers, or create a new headers object.
   const headers = requestInit?.headers
     ? (requestInit.headers as Headers)
     : new Headers();
-  // If we have the response object, add the debug header.
-  if (res) {
+
+  // If we have the response object, add the debug header for fetch requests
+  // and set appropriate cache-control headers on the active response.
+  if (res && typeof res !== 'boolean') {
+    // Ensure api response contains surrogate key headers.
     headers.set('Pantheon-Debug', '1');
+
+    res.setHeader('Cache-Control', cacheControl);
   }
+
   // Set the updated headers for fetch.
   requestInit.headers = headers;
 
@@ -36,13 +47,6 @@ const defaultFetch = (
       if (res && response.headers.has('Surrogate-Key-Raw')) {
         addSurrogateKeyHeader(
           response.headers.get('Surrogate-Key-Raw'),
-          res as ServerResponse
-        );
-      }
-
-      if ((res as ServerResponse) && response.headers.has('Cache-Control')) {
-        updateMaxAge(
-          response.headers.get('Cache-Control') as string,
           res as ServerResponse
         );
       }