+++ packages/wordpress-kit/vite.config.js
@@ -4,7 +4,7 @@ module.exports = {
   build: {
     lib: {
       entry: path.resolve(__dirname, 'index.ts'),
-      name: 'WordpressKit',
+      name: 'WordPressKit',
       fileName: format => `wordpress-kit.${format}.js`,
     },
   },