+++ starters/next-drupal-starter/pages/recipes/index.jsx
@@ -1,5 +1,6 @@
 import { useRouter } from "next/router";
 import { NextSeo } from "next-seo";
+import { sortDate } from "@pantheon-systems/nextjs-kit";
 import { isMultiLanguage } from "../../lib/isMultiLanguage";
 import {
   getCurrentLocaleStore,
@@ -13,7 +14,7 @@ import PageHeader from "../../components/page-header";
 // This file can safely be removed if the Drupal
 // instance is not sourcing Umami data
 export default function RecipeListTemplate({
-  recipes,
+  sortedRecipes,
   footerMenu,
   hrefLang,
   multiLanguage,
@@ -31,7 +32,7 @@ export default function RecipeListTemplate({
       <PageHeader title="Recipes" />
       <section>
         <RecipeGrid
-          data={recipes}
+          data={sortedRecipes}
           contentType="recipes"
           multiLanguage={multiLanguage}
           locale={locale}
@@ -70,9 +71,21 @@ export async function getServerSideProps(context) {
       res: context.res,
     });
 
+    if (!recipes) {
+      throw new Error(
+        "No recipes returned. Make sure the objectName and params are valid!"
+      );
+    }
+
+    const sortedRecipes = sortDate({
+      data: recipes,
+      key: "created",
+      direction: "desc",
+    });
+
     return {
       props: {
-        recipes,
+        sortedRecipes,
         footerMenu,
         hrefLang,
         multiLanguage,